Lint usage of dracut-lib function.
This call site seems to have treated strstr as if the search term
would be treated as a pattern rather than a string. As it happened,
before these patches, the missing quotes in the definition of strstr
meant it /would/ treat the search string as a pattern, raising the
possibility that this call site depended on that behavior. It didn't,
though, because under the old behavior the extra * here caused no
problem but also had no effect. Now, it would cause a problem, so it's
out. Since strstr's namesake is a C library function that's known
to search only for literal strings, having it behave differently in
dracut-lib would seem to be asking for surprises.